+++ srcb/ft_printf_bonus.h
@@ -6,7 +6,7 @@
 /*   By: passunca <passunca@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/31 20:41:49 by passunca          #+#    #+#             */
-/*   Updated: 2024/02/11 19:19:31 by passunca         ###   ########.fr       */
+/*   Updated: 2024/02/11 21:32:08 by passunca         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -63,7 +63,7 @@ int			ft_print_c(char c, t_format *prsd);
 int			ft_print_s(const char *str, t_format prsd);
 int			ft_print_str(const char *str);
 int			ft_print_di(int n, t_format *p);
-int			ft_print_f(int n, t_format *p);
+int			ft_print_f(double n, t_format *p);
 int			ft_print_u(unsigned n, t_format prsd);
 int			ft_print_p(unsigned long int n, t_format prsd);
 int			ft_print_hex(t_format prsd, va_list ap);
